pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "binance-exporter"
version = "0.1.0"
description = "Prometheus exporter for Binance flexible crypto loans"
requires-python = ">=3.11"
license = { text = "MIT" }
dependencies = [
    "prometheus-client>=0.21.0",
    "python-dotenv>=1.0.0",
    "backoff>=2.2.1",
    "binance-sdk-crypto-loan>=4.3.0",
    "binance-common>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
]

[project.scripts]
binance-exporter = "binance_exporter.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--tb=short"
